# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  output.BuscarClienteOutput:
    properties:
      email:
        type: string
      id:
        type: string
      nome:
        type: string
      role:
        type: string
      telefone:
        type: string
    type: object
...
  output.CatalogoOutput:
    properties:
      categoria:
        type: string
      duracaoPadrao:
        type: integer
      id:
        type: string
      imagemUrl:
        type: string
      nome:
        type: string
      preco:
        type: integer
    type: object
...
  request.ClienteRequest:
    properties:
      email:
        example: joao@email.com
        type: string
      nome:
        example: João da Silva
        maxLength: 100
        minLength: 3
        type: string
      senha:
        example: senha123
        minLength: 8
        type: string
      telefone:
        example: "62999677481"
        maxLength: 15
        minLength: 8
        type: string
    required:
    - email
    - nome
    - senha
    - telefone
    type: object
...
  request_catalogo.CatalogoRequest:
    properties:
      categoria:
        example: Redes
        maxLength: 50
        minLength: 3
        type: string
      duracao_padrao:
        example: 20
        type: integer
      image_url:
        example: https://tdfuderuzpylkctxbysu.supabase.co/storage/v1/object/public/imagens/bb515383d2f6ef76.jpg
        type: string
      nome:
        example: Tecnico de Redes
        maxLength: 100
        minLength: 3
        type: string
      preco:
        example: 10000
        type: integer
    required:
    - categoria
    - duracao_padrao
    - image_url
    - nome
    - preco
    type: object
  request_catalogo.CatalogoUpdateRequest:
    properties:
      categoria:
        example: Estética Facial
        maxLength: 50
        minLength: 3
        type: string
      duracao_padrao:
        example: 40
        type: integer
      image_url:
        example: https://tdfuderuzpylkctxbysu.supabase.co/storage/v1/object/public/imagens/bb515383d2f6ef76.jpg
        type: string
      nome:
        example: Limpesa de Peele
        maxLength: 100
        minLength: 3
        type: string
      preco:
        example: 10000
        type: integer
    required:
    - categoria
    - duracao_padrao
    - image_url
    - nome
    - preco
    type: object
...
  response_catalogo.CatalogoListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/output.CatalogoOutput'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response_catalogo.CatalogoResponse:
    properties:
      categoria:
        type: string
      duracao_padrao:
        type: integer
      id:
        type: string
      image_url:
        type: string
      nome:
        type: string
      preco:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API para gestão de clientes e serviços com autenticação JWT.

    Credenciais do Administrador (para testes):
    Email: admin@admin.com
    Senha: admin123
  title: API de Agendamentos
  version: "1.0"
paths:
...
  /catalogos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de catálogos, com page e limit para paginação.
      parameters:
      - default: 1
        description: Número da página
        in: query
        name: page
        type: integer
      - default: 10
        description: Quantidade de itens por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response_catalogo.CatalogoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Lista todos os catálogos com paginação
      tags:
      - Catalogos
    post:
      consumes:
      - application/json
      description: Cadastra um serviço que pode ser oferecido por um prestador. Requer
        autenticação JWT (APENAS Admin).
      parameters:
      - description: Dados do Catálogo
        in: body
        name: catalogo
        required: true
        schema:
          $ref: '#/definitions/request_catalogo.CatalogoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Catálogo criado com sucesso
          schema:
            $ref: '#/definitions/response_catalogo.CatalogoResponse'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - apenas Admin pode criar catálogos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Prestador não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Catálogo já existente
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cria um novo catálogo de serviços
      tags:
      - Catalogos
  /catalogos/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um catálogo pelo ID. Requer autenticação JWT (APENAS Admin).
      parameters:
      - description: ID do Catálogo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Catálogo deletado com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - apenas Admin pode deletar catálogos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Catálogo não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deleta um catálogo existente
      tags:
      - Catalogos
    get:
      consumes:
      - application/json
      description: Retorna os dados de um catálogo específico usando seu ID.
      parameters:
      - description: ID do Catálogo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Catálogo encontrado com sucesso
          schema:
            $ref: '#/definitions/response_catalogo.CatalogoResponse'
        "400":
          description: ID inválido fornecido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Catálogo não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor ou falha de infraestrutura
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Busca um catálogo pelo ID
      tags:
      - Catalogos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um catálogo pelo ID. Requer autenticação JWT
        (APENAS Admin).
      parameters:
      - description: ID do Catálogo
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados do Catálogo
        in: body
        name: catalogo
        required: true
        schema:
          $ref: '#/definitions/request_catalogo.CatalogoUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Catálogo atualizado com sucesso
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - apenas Admin pode atualizar catálogos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Catálogo não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualiza um catálogo existente
      tags:
      - Catalogos
  /clientes:
    post:
      consumes:
      - application/json
      description: Recebe dados de nome, email, telefone e senha para registrar um
        novo cliente.
      parameters:
      - description: Dados do Cliente
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/request.ClienteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cliente criado com sucesso
          schema:
            $ref: '#/definitions/output.BuscarClienteOutput'
        "400":
          description: Dados inválidos (erro de validação do binding)
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Cliente já cadastrado (Email já existe)
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Falha na persistência de dados ou erro interno
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Cadastra um novo cliente
      tags:
      - Clientes
  /clientes/{id}:
    get:
      consumes:
      - application/json
      description: Retorna os dados de um cliente específico usando seu ID. Requer
        autenticação JWT.
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cliente encontrado com sucesso
          schema:
            $ref: '#/definitions/output.BuscarClienteOutput'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - você só pode acessar seus próprios dados
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Cliente não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor ou falha de infraestrutura
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Busca um cliente pelo ID
      tags:
      - Clientes
...
securityDefinitions:
  BearerAuth:
    description: Digite "Bearer " seguido do token JWT recebido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"