# backend/internal/adapters/http/catalogo/request_catalogo/request_cadastro.go
package request_catalogo

import (
	"meu-servico-agenda/internal/core/application/input"

	"github.com/rs/xid"
)

type CatalogoRequest struct {
	Nome          string `json:"nome" binding:"required,min=3,max=100" example:"Tecnico de Redes" swagger:"desc('Nome do serviço')"`
	DuracaoPadrao int    `json:"duracao_padrao" binding:"required" example:"20" swagger:"desc('Duração padrão do serviço em minutos')"`
	Preco         int    `json:"preco" binding:"required" example:"10000" swagger:"desc('Preço do serviço em centavos')"`
	ImagemUrl     string `json:"image_url" binding:"required,url" example:"https://tdfuderuzpylkctxbysu.supabase.co/storage/v1/object/public/imagens/bb515383d2f6ef76.jpg"`
	Categoria     string `json:"categoria" binding:"required,min=3,max=50" example:"Redes" swagger:"desc('Categoria do serviço')"`
}

func (cr *CatalogoRequest) ToCatalogoInput() *input.CatalogoInput {
	return &input.CatalogoInput{
		ID:            xid.New().String(),
		Nome:          cr.Nome,
		DuracaoPadrao: cr.DuracaoPadrao,
		Preco:         cr.Preco,
		Categoria:     cr.Categoria,
		ImagemUrl:     cr.ImagemUrl,
	}
}